requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
duckdb==1.0.0
idna==3.7
python-dotenv==1.0.1
requests==2.32.3
ruff==0.4.10
urllib3==2.2.2
